refactored the repo
Refactored the entire repository based on languages used and concept undertaken.
