+++ useful_competitive_codes/23766503.cpp
